api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.2",
    "cryptography>=44.0.2",
    "fastapi[standard]>=0.112.2",
    "pydantic-settings>=2.9.1",
    "pymysql>=1.1.1",
    "pytz>=2025.2",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
hello = "uv_docker_example:hello"

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
dev-dependencies = [
    "ruff>=0.6.2",
    "fastapi-cli>=0.0.5",
    "requests>=2.32.3",
    "redis>=5.2.1",
    "celery>=5.5.2",
]
